some fixes

+++ src/appointments/models/appointment.ts
@@ -1,5 +1,5 @@
 import { Place } from './place';
-import { ObjectType, Field, Int } from '@nestjs/graphql';
+import { ObjectType, Field, Int, Float } from '@nestjs/graphql';
 import { Patient } from './patient';
 
 @ObjectType()
@@ -10,13 +10,13 @@ export class Appointment {
   @Field()
   state: string;
 
-  @Field((type) => Int)
+  @Field((type) => Float)
   createdAt: number;
 
   @Field((type) => Patient)
   patient: Patient;
 
-  @Field((type) => Int, { nullable: true })
+  @Field((type) => Float, { nullable: true })
   date?: number;
 
   @Field((type) => Place, { nullable: true })
